requirements.txt
msal
requests
python-dotenv
openai